@@ -6,14 +6,8 @@ import { HttpModule } from '@angular/http';
 
 import { AppComponent } from './app.component';
 
-import 'rxjs/add/observable/interval';
-import 'rxjs/add/observable/range';
-import 'rxjs/add/operator/skip';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/take';
-import 'rxjs/add/operator/switch';
-import 'rxjs/add/operator/do';
+import { Observable, interval, range, pipe } from 'rxjs';
+import { switchMap, map, take, skip, switchAll, tap } from 'rxjs/operators';
 
 @NgModule({
   declarations: [
